server/requirements.txt
alembic==1.0.7
APScheduler==3.5.3
fakeredis==0.16.*
gino==0.8.1
psycopg2==2.7.7
pyjwt==1.7.*
python-dotenv==0.10.1
redis==3.0.*
testing.postgresql==1.3.0
tornado==5.1.*
